Fixed source code to prevent POSIX warnings/deprecate notifications...
Fixed all POSIX warnings/deprecation notifiations

Fixed source code issues where possible loss of data would make compiler
display precision loss warnings.

Replaced qboolean enum values, {true;false} with {qtrue;qfalse} (to
avoid ambigity).

+++ public/pman_triangleffect.h
@@ -193,7 +193,7 @@ protected:
 		// This check will help prevent a class frome being defined later,
         //  that is larger than the max size MemoryPool is expecting,
         //  from being successfully allocated.
-        if (size > (unsigned long) CMiniMem::Instance()->MaxBlockSize())
+        if (size > (int) CMiniMem::Instance()->MaxBlockSize())
 		{
 #ifdef _WIN32
             throw "New particle class is larger than memory pool max size, update lMaxParticleClassSize() function.";
